Refactor GNN Model Pipeline: Modularization and Architecture Updates (#10)
### Description of Changes
This pull request includes significant improvements and refactoring to
the Graph Neural Network (GNN) pipeline in the NOAA-EMC/ocelot
repository. The changes include modularizing the `EndToEndGNN.py` file
into separate components and updating the `README.md` to reflect the new
structure. Below are the key changes:
The associated issue is #9 
**1. Modularization:**

- Split the `EndToEndGNN.py` file into several modular files within the
`gnn_model` directory.

- `gnn_model.py`: Contains the GNNLightning model (encoder, processor,
decoder).
- `gnn_datamodule.py`: Handles graph construction and PyTorch Lightning
DataModule.
  - `train_gnn.py`: Main training script.
  - `obs_to_mesh.py`: Manages cutoff-based encoder edges.
  - `mesh_to_mesh.py`: Manages multi-scale mesh processor edges.
  - `mesh_to_target.py`: Manages KNN-based decoder edges.
  - `mesh_creation.py`: Manages mesh and graph construction.
- `process_timeseries.py`: Handles Zarr binning and feature extraction.
  - `timing_utils.py`: Contains the resource-logging decorator.

**2. Updates to README.md:**

- Updated the README.md to reflect the new modular structure.
- Detailed the model architecture, including the encoder, processor, and
decoder components.

**3. Key Improvements:**

  - Improved mesh generation to track and utilize multilevel edges.
- Enhanced edge creation to support optional geodesic distance
attributes.
- Refactored the GNN model to incorporate MLP-based encoding and
decoding with distance-based features and weighted aggregation.

These updates aim to enhance the performance and flexibility of the
graph construction and GNN modeling processes.

---------
