+++ gnn_model/mesh_to_target.py
@@ -1,0 +1,83 @@
+import torch
+from sklearn.neighbors import NearestNeighbors
+
+
+class MeshTargetKNNConnector:
+    """
+    Computes K-Nearest Neighbors (KNN)-based edges for decoding
+    (connecting hidden mesh nodes to target data nodes).
+
+    Attributes:
+        num_nearest_neighbours (int): Number of nearest neighbors to consider for each target node.
+
+    Methods:
+        add_edges(mesh_graph, target_latlon_rad, mesh_latlon_rad):
+            Connects each target observation location to its K-nearest mesh nodes.
+
+        create_edge_index(edge_list, edge_weights):
+            Converts edge list to PyTorch Geometric `edge_index` format with associated edge weights.
+    """
+
+    def __init__(self, num_nearest_neighbours: int):
+        """
+        Initializes the MeshTargetKNNConnector class.
+
+        Parameters:
+            num_nearest_neighbours (int): Number of nearest neighbors to connect each target node to.
+
+        Raises:
+            AssertionError: If `num_nearest_neighbours` is not a positive integer.
+        """
+        assert isinstance(num_nearest_neighbours, int), "num_nearest_neighbours must be an integer"
+        assert num_nearest_neighbours > 0, "num_nearest_neighbours must be positive"
+        self.num_nearest_neighbours = num_nearest_neighbours
+
+    def add_edges(self, mesh_graph, target_latlon_rad, mesh_latlon_rad):
+        """
+        Connects each target observation node to its K-nearest mesh nodes using the Haversine metric.
+
+        Parameters:
+            mesh_graph (networkx.Graph): The mesh graph containing node connectivity.
+            target_latlon_rad (numpy.ndarray): Array of shape (M, 2) containing target node
+                                               coordinates (latitude, longitude in radians).
+            mesh_latlon_rad (numpy.ndarray): Array of shape (N, 2) containing mesh node
+                                             coordinates (latitude, longitude in radians).
+
+        Returns:
+            tuple:
+                - torch.Tensor: Edge index tensor of shape (2, E), where each column represents
+                                an edge from a mesh node to a target node.
+                - torch.Tensor: Edge attribute tensor containing distance weights for each edge.
+        """
+        knn = NearestNeighbors(n_neighbors=self.num_nearest_neighbours, metric="haversine")
+        knn.fit(mesh_latlon_rad)
+
+        distances, indices = knn.kneighbors(target_latlon_rad)
+
+        edge_list = []
+        edge_weights = []
+
+        # Create directed edges from mesh nodes to target nodes
+        for target_idx, mesh_neighbors in enumerate(indices):
+            for i, neighbor in enumerate(mesh_neighbors):
+                edge_list.append([neighbor, target_idx])  # Mesh node → Target node
+                edge_weights.append(distances[target_idx, i])  # Distance weight
+
+        return self.create_edge_index(edge_list, edge_weights)
+
+    def create_edge_index(self, edge_list, edge_weights):
+        """
+        Converts an edge list to PyTorch Geometric `edge_index` format with edge attributes.
+
+        Parameters:
+            edge_list (list): List of edges, where each entry is a pair [source, target].
+            edge_weights (list): List of distance weights corresponding to the edges.
+
+        Returns:
+            tuple:
+                - torch.Tensor: Edge index tensor (2, E) in COO format.
+                - torch.Tensor: Edge attribute tensor containing distance weights.
+        """
+        edge_index = torch.tensor(edge_list, dtype=torch.long).t().contiguous()
+        edge_attr = torch.tensor(edge_weights, dtype=torch.float32)
+        return edge_index, edge_attr
